FDG-6273 Refactor national debt explained
