+++ src/layouts/explainer/sections/national-debt/national-debt-explained/national-debt-explained.module.scss.d.ts
@@ -63,49 +63,9 @@ export const paleBg: string;
 export const debtExplainerPrimary: string;
 export const debtExplainerSecondary: string;
 export const debtExplainerLightSecondary: string;
-export const keyTakeawaysContent: string;
-export const noMarginBottom: string;
-export const iconBackground: string;
-export const icon: string;
-export const offsetIcon: string;
-export const fundingProgramAccordion: string;
-export const spendingCategoriesAccordionContent: string;
-export const spendingCategoriesTable: string;
-export const row: string;
-export const firstColumn: string;
-export const secondColumn: string;
-export const growingNationalDebtSectionAccordion: string;
-export const accordionHeader: string;
-export const rectangle: string;
-export const accordionTable: string;
-export const accordionFooter: string;
-export const debtBreakdownSectionGraphContainer: string;
-export const title: string;
-export const simple: string;
-export const header: string;
-export const DebtOverLast100yContainer: string;
-export const growingNationalDebtSectionGraphContainer: string;
-export const debtTrendsOverTimeSectionGraphContainer: string;
-export const titleBreakdown: string;
-export const subTitle: string;
-export const headerContainer: string;
-export const subHeader: string;
-export const growingNationalDebtSectionGraph: string;
-export const footerContainer: string;
-export const barChartContainer: string;
-export const multichartContainer: string;
-export const aveInterestLegend: string;
-export const multichartLegend: string;
-export const debtLegend: string;
-export const lineChartContainer: string;
-export const postGraphContent: string;
-export const postGraphAccordionContainer: string;
-export const diveDeeperQuoteRight: string;
-export const diveDeeperQuote: string;
-export const diveDeeperQuoteLeft: string;
-export const diveDeeperCitation: string;
-export const diveDeeperLink: string;
-export const debtAccordion: string;
-export const debtCeilingAccordion: string;
-export const growingNationalDebt: string;
-export const multichartWrapper: string;
+export const nationalDebtExplainedTextContent: string;
+export const nationalDebtExplainedTable: string;
+export const tableIcon: string;
+export const borderBottom: string;
+export const paddingTop: string;
+export const diveDeeperContent: string;
