+++ src/pages/americas-finance-guide/afg-overview.module.scss.d.ts
@@ -80,10 +80,6 @@ export const spendingCategoriesTable: string;
 export const row: string;
 export const firstColumn: string;
 export const secondColumn: string;
-export const nationalDebtExplainedTextContent: string;
-export const nationalDebtExplainedTable: string;
-export const tableIcon: string;
-export const borderBottom: string;
 export const growingNationalDebtSectionAccordion: string;
 export const accordionHeader: string;
 export const rectangle: string;
@@ -118,8 +114,6 @@ export const diveDeeperLink: string;
 export const debtAccordion: string;
 export const debtCeilingAccordion: string;
 export const growingNationalDebt: string;
-export const paddingTop: string;
-export const diveDeeperContent: string;
 export const multichartWrapper: string;
 export const deficitAccordion: string;
 export const mainContainer: string;
